Add program info popup
There is no mention of the program version or author on the calculator's side.

This adds a popup that will display the attached message upon the first
program load. Note: the detection depends on the program being stored in RAM;
it will always display if the program was archived before the initial run.

The about popup can also be opened by running the program with "about" as the
only argument. This does not reset the first program load flag.

The following is the message displayed on the device (wrapped like the TI-89
displays it):

> chip8-ti68k v0.9
> A (S)CHIP-8 emulator for ti68k
> graphing calculators.
>
> Copyright 2022 Peter Lafreniere
> This is free software. See COPYING for
> more details.