Fix calling program with argument
The program is supposed to skip the rom selection dialog when there is a
string argument passed. It is also supposed to display the about popup when
the argument passed is "about".

This functionality was broken. It only opened the about popup regardless of
the passed argument. This patch reverts to the previous behavior, but retains
the about menu parameter.