Add readme
This project is close enough to a full release that not having any
documentation has become an issue. Add the first version of a README file.

It is unlikely that this version of the file is final. Despite that, this
commit marks the end of code work for the v1.0 release. All changes until then
will be documentation related.