Add plane select (fn01) XO-CHIP opcode
With access to drawing with planes, it's simple to allow programs to set
the plane bitmask. This enables 4-colour graphics on ti68k calculators.

Tested with the title screen of OctoPartyMix.ch8

Works well, but not very useful on its own. Next up are dummy handlers
for XO-CHIP sound opcodes, the scroll up opcode, and long pointers
(64KiB address space).