Use TI-OS error routines
Currently we return the error code after calling each function in
opcodes.c. This is unoptimal, as the majority of the time the result is
E_OK. So switch that file to use the native exception handling routines.

This change helps the integration of XO-CHIP extended memory, as well as
shaving off (on average) 60 bytes of binary size (~0.7%).